wip: move routes to root
